Renamed "Home" template to "KanbanBoard"
