+++ src/templates/KanbanBoard/__mocks__/clearCookies.ts
