+++ src/templates/KanbanBoard/clearCookies.ts
