+++ src/templates/KanbanBoard/index.tsx
@@ -3,21 +3,25 @@
 import { StateDTO } from "@/model/State";
 import { TaskDTO } from "@/model/Task";
 import { UserDTO } from "@/model/User";
-import { TaskList } from "@/templates/Home/TaskList";
 import { theme } from "@/templates/theme";
 import { ChakraProvider } from "@chakra-ui/react";
 import { useEffect } from "react";
 import { DndProvider } from "react-dnd";
 import { HTML5Backend } from "react-dnd-html5-backend";
+import { Layout } from "./Layout";
 import { useZustand } from "./model";
 
-export interface HomeProps {
+export interface KanbanProps {
   initialStates: StateDTO[];
   initialTasks: TaskDTO[];
   initialUser?: UserDTO;
 }
 
-export function Home({ initialStates, initialTasks, initialUser }: HomeProps) {
+export function KanbanBoard({
+  initialStates,
+  initialTasks,
+  initialUser,
+}: KanbanProps) {
   const initialize = useZustand((store) => store.initialize);
 
   // Initialize zustand with the server-side data
@@ -28,7 +32,7 @@ export function Home({ initialStates, initialTasks, initialUser }: HomeProps) {
   return (
     <ChakraProvider theme={theme}>
       <DndProvider backend={HTML5Backend}>
-        <TaskList />
+        <Layout />
       </DndProvider>
     </ChakraProvider>
   );
