+++ src/templates/KanbanBoard/model/types.ts
